requirements.txt
pygame==2.5.2
openai==1.65.4
anthropic==0.19.0
python-dotenv==1.0.1
pydantic==2.6.3
pytest==8.0.2
pandas==2.2.1
matplotlib==3.8.3
rich==13.7.0
httpx==0.28.1
tqdm==4.67.1 